requirements.txt
pytorch
transformers
parsivar
advertools